pyproject.toml
[tool.poetry]
name = "ScandEval"
version = "8.0.0"
description = "Evaluation of pretrained language models on mono- or multilingual Scandinavian language tasks."
authors = ["Dan Saattrup Nielsen <saattrupdan@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://scandeval.github.io"
repository = "https://github.com/saattrupdan/ScandEval"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
tqdm = "^4.62.0"
huggingface-hub = ">=0.7.0,<1.0.0"
datasets = "^2.7.0"
click = "^8.1.3"
termcolor = "^2.0.0"
numpy = "^1.23.0"
sentencepiece = ">=0.1.96,<1.0.0"
protobuf = ">=3.20.0,<3.21.0"
seqeval = "^1.2.2"
pandas = "^1.4.0"
python-dotenv = ">=0.20.0,<1.0.0"
sacremoses = ">=0.0.53,<1.0.0"
jax = ">=0.4.1,<1.0.0"
flax = ">=0.6.3,<1.0.0"
jaxlib = ">=0.4.1,<1.0.0"
pyinfer = ">=0.0.3,<1.0.0"
openai = ">=0.27.8,<1.0.0"
tiktoken = ">=0.4.0,<1.0.0"
levenshtein = ">=0.21.1,<1.0.0"
einops = ">=0.6.0,<1.0.0"
bitsandbytes = {markers="sys_platform != 'darwin' or platform_machine != 'arm64'", version = ">=0.40.0.post4,<1.0.0"}
requests = "^2.31.0"
peft = ">=0.4.0,<1.0.0"
pydantic = "^2.3.0"
torch = "^2.1.1"
transformers = "^4.34.0"
accelerate = ">=0.23.0,<1.0.0"
evaluate = ">=0.4.0,<1.0.0"
nltk = "^3.8.1"
rouge-score = ">=0.1.2,<1.0.0"
bert-score = ">=0.3.13,<1.0.0"
sacrebleu = "^2.3.2"
cer = "^1.2.0"

[tool.poetry.group.dev.dependencies]
pdoc = "^14.0.0"
pytest = "^7.4.0"
pre-commit = "^3.3.3"
black = "^23.7.0"
lxml = "^4.9.3"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
readme-coverage-badger = ">=0.1.2,<1.0.0"

[tool.poetry.scripts]
scandeval = "scandeval.cli:benchmark"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    '--verbose',
    '--durations=10',
    '--color=yes',
    '-s',
    '-vv',
    '--doctest-modules',
    '--cov=src/scandeval',
]
xfail_strict = true
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::ImportWarning",
]
log_cli_level = "info"
testpaths = ["tests", "src/scandeval"]

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''

[tool.ruff]
target-version = "py311"

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
